optimfg-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
# psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
numpy==1.26.4
pymoo==0.6.1.1
